backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
pillow
djangorestframework-simplejwt
django-filter
gunicorn>=21.2.0
whitenoise>=6.6.0
